fix: complex/maximal-payload live verification (310/317 elements confirmed)
Exercised every exposed config element + every polymorphic variant live; fixes:
- x509-ca: QC_TYPES enum (WEB/ESIGN/ESEAL/NONE), SHA3 underscore casing, eTSITransactionLimit.valueLimit int
- x509-certificate: typed AIA/POLICY/EKU(custom required)/empty_extensions schemas for enroll overrides
- rbac: x509 credential keyPair retain-on-update; credentials onCredentialsExpiration triggers
- revocation: assign_ocsp_signer_to_ca CA strip set keeps required privateKey + dn
Variants confirmed live: CA managed/external/hybrid; keystore software/pkcs11/aws/akv/gcp;
trigger email/rest(all auth types); credential/identity-provider all types; template all
KeyUsage/EKU/SAN/DN/extension enums; 15 key algorithms; 9 hash algorithms; SSH user/host.
283 unit tests + live e2e green.

+++ src/tools/x509-ca/schema.ts
@@ -103,7 +103,9 @@ const qcStatementSchema = z
     eTSIPDS: z.record(z.string(), z.string()).optional(),
     eTSITransactionLimit: z
       .object({
-        valueLimit: z.number(),
+        // The server takes integers for both (error.expected.int on a
+        // float). currencyCode must be 3 uppercase chars (server-enforced).
+        valueLimit: z.number().int(),
         valueLimitExp: z.number().int(),
         currencyCode: z.string(),
       })
